Markasdup lib removed; markasdup CLI explanation improved
